chore: remove coverage dependency on Events.ts
Another step towards removing Events.ts. This commit updates the
coverage utils to instead parse each module and check for the presence
of an `XEmittedEvents` object which indicates that module has events
that we need to check coverage on.

I also refactored how we store the list of what we consider public API
to make it easier to use in coverage and then in the
`installAsyncStackHelper` that gets run in `initialize.ts`.