chore: remove redundant run config for depend-prah
Remove the redundant rule for dependency graph as the `src/index.ts` is
not required. This was introduced in 64c9c709, but the filestructure has
changed from since then and it doesn't work as `src/index.ts` is
replaced with `src/node.ts` and `src/web.ts` as per the use case.