chore: shorter node checker function
`isNode` checks for the node version in which the process is running and
it checks for the node key inside the versions. This will be exactly
same to as using `process.version` as they are set in the native code
and `process.version` is just a shorthand for `process.versions.node`.