Running Puppeteer in an EC2 instance with Amazon-Linux
Dear Puppeteer Founders, Developers in general,

Thank you for creating such an wonderful tool which enables people to easily perform E2E acceptance tests with out any huzzle. I recently encountered a huge problem in running Puppeteer because puppeteer was not able to spin up chromium instance in an EC2 instance with amazon-linux and I found a way to troubleshoot that, felt like adding this to troubleshooting guide. 

I hope this helps others who are consuming it.

Thank you