Remove broken executable "resolver" for arm64 arch
There was never any reasonable excuse for this branch to exist. There is no arch where a hardcoded executable path should be used in preference to the path resolver, as it simply prevents you from using PUPPETEER_EXECUTABLE_PATH, which is the preferred option for not hardcoding the executable path in usage. Removing this also provides better error handling, as the resolver provides an error message, whereas this broken behaviour just fails.