chore: set `module` to `esnext` in `tsconfig.json`
The main `tsconfig.json` file is only used for API Extractor, and by
VSCode to provide type information. It is _not_ used to compile
Puppeteer for shipping. Therefore we can specify `module: "esnext"` in
here so that VSCode knows we can use all the latest and greatest
module features (primarily, dynamic imports). In `tsconfig.cjs.json`
and `tsconfig.esm.json` we set the `module` setting for CJS/ESM
respectively.