chore(docs): ensure old docs are removed when generating
I lost time today due to some old docs files lingering for code that
is now gone. To avoid that happening, let's remove the directories
before generating.

As part of this work I also changed the API Extractor to not output to
`temp/X.api.json` and instead `docs-api-json/X.api.json` to make it
clearer what that folder is for.