docs: add warning about version of try-puppeteer.appspot.com
The URL https://try-puppeteer.appspot.com/ is mentioned around the top of `README.md` even if it is running a very outdated version (v1.4.0). I suggest either (1) remove it from the docs or (2) putting a warning that it is a bit outdated.  
I made a suggestion for the latter in this PR.