chore: reduce docusaurus versions that we generate
The Docusaurus process has been OOMing on the bots; this is a problem
we'll want to fix properly when we are shipping the new docs fully, as
we want to generate docs for old versions; but for now whilst we're not
fully pushing users to this site, we can just generate one version and
not worry about multiples.

This commit also bumps the version of Docusaurus to the latest beta
version.