puppeteer/SECURITY.md
This needs to go Private! I need access but do not want it linked into my web versions!I want the owner of github to call me as soon as possible! The contract that paid for this job is a fradulent contract! Joseph Alexander Samudio did not have authorization to assign a contract for this! Neither did John Smith!
Please link to my account NU7891 in timeout mode! With permission to run command by the owner of GITHUB only! SET AS PRIVATE ON MY ACCOUNT! PLEASE! 2105024220