chore: support WebDriver BiDi browser instances
This PR adds a basic support for WebDriver BiDi that
currently includes only the ability to establish a connection
and shutdown the browser. Therefore, the implementation is marked
as internal and won't show up in the changelog as it's barely useful
at the moment.

The API classes are kept as classes instead of interfaces so
that clients relying on instanceof checks still work. Also, we
disable noUnusedParameters for now as those API classes need unused
variable names for documentation generation.