chore: include component in puppeteer tag (#9303)
Removes `include-component-in-tag` from the release please config which
means that Puppeteer package releases will now be following the format
`puppeteer-vX.X.X`. Also, this PR updates the docker publishing action
to correctly parse such a tag to extract the `X.X.X` part, that is to be
used as the docker image label.