chore: fix third_party build
The third_party folder is compiled with tsc and then rolled up in-place.
This means that the next build might try to compile the file that
has been rolled up leading to build errors. It seems the safest solution
is to always rebuild third_party folder as the size is relatively small.