fix: implement protocol-level timeouts
This PR implements timeouts for the protocol-level calls and a new launch/connect option called protocolTimeout. The default value will be 30s and there is an option to disable it by providing an undefined/negative/0 value.

Drive-by: refactored callbacks to encapsulate timeouts and reduce duplication. Also, marked ConnectionCallback as internal as there was no way to use it externally anyway.

Closes #9801

For follow-ups: we should probably mark connection as an internal class (but we need to deprecate it first and see if there are users for it) and changes the constructor parameters to accept an object instead of positional arguments.