chore: skip building on install
Currently, we try to compile the installation script during
`npm i` in the development repository. This is a fragile process
because the build might not succeed during the installation (missing packages)
and it adds to the installation time. This PR removes the build step and
instead recommends running the install after a build.