fix: increase the default protocol timeout
This PR increases the protocol timeout to be less likely triggered
by normal operations but still indicate errors in case of backend
issues instead of waiting forever. Unfortunately, the current
script evaluation API does not allow configuring the timeouts per
operations and it's not possible to change this due to variadic arguments
accepted in evaluate and evaluateHandle. We could consider exposing
new methods which also accept a timeout but for now let's see if increasing
the connection timeout is good enough.

Issue #9927