ci: separate unit tests
Unit tests are moved next to the code they test.
`npm run unit` is the command to run them. The test files
are excluded from the publishing and an eslint rule is added
to prevent regular code from depending on the test files.

`.test.js` is chosen because it is the default suffix supported
by the Node test runner. Running unit tests required Node v20+.