fix: stop relying on vm2 (via proxy agent)
Originally, we hoped to support all proxy kinds by integrating
the proxy-agent module. It turns out that pac file proxy
implementation relies on the vm2 module that recently became
deprecated due to security problems. The recommended replacement
requires a compiler installed which would lead to extra friction
for many users. It looks like the best would be to drop the pac
proxy support (which we never announced publicly) and stop
relying on the proxy-agent module that has vm2 as a dependency.