feat: support launch aliases
With this PR, @puppeteer/browsers starts maintaining a local
mapping from an alias to the locally installed version.
This allows using aliases such as `canary`/`latest` when launching
the browser.

Note that the launch command would only consult metadata files
and won't attempt to re-check aliases by contacting the server.