fix: ignore frames removed by PDF viewer
PDF viewer in Chrome replaces the inner iframes
with an extension web view. The iframes are still
observable in Puppeteer up until the point when
they get replaces. In this case, Chrome sends
the frameDetached event with the reason swap.

Unlike the regular swap where a frame is replace with
an OOPIF or vice versa, with PDF viewer the frame is
just gone. This PR works around this issue by
ignoring any frames that are being swapped.