fix: default to RAF polling if visible||hidden is set
The regression is caused by the visible flag no longer
changing the polling type to RAF. This solves the
case reported in the related bug when changes within
shadow roots are tracked with visible=true. The underlying
issue is that MutationPoller does not consider shadow
roots at all. There is a related spec issue to support
shadow roots in MutationObserver but as long as it is
not supported we need a client side solution.

This PR adds a test for that case as well as a test
when visibility is changed without DOM mutations.

Refs: #13152
Related: https://github.com/whatwg/dom/issues/1287
Drive-by: adds a test for mutations inside shadow roots.