fix: use correct paper format units
When format is set, we didn't convert the internal values (in inches) to the expected measuring system, for Firefox and WebDriver BiDi the expected unit is centimeters, but we were providing the value in inches.

Now, all supported paper format have defined sizes in both centimeters and inches and when PDF options are parsed, we will returned size of the selected format in provided unit.

Fix #13515