feat: expose utilities for managing browser artifacts manually
When integrating Puppeteer into e.g. Bazel, we cannot rely on Puppeteer
downloading browsers for users. Instead, ruleset authors compute a set
of download URLs, SHAs etc so that Bazel can perform the binary
download.

In order to be able to use Puppeteer-core at runtime though, or
generally stay up-to-date, we want to leverage `@puppeteer/browsers` to
manage browser versions and URLs. This commit expands the API surface to
expose the following things:

- `getDownloadURLs` for retrieving download URLs
- `BrowserTag` so that `resolveBuildId` can rely on it in a type-safe way.
- `computeExecutablePath` is modified to also work without a `cacheDir`.