Update package.json to fix security hole
Pinned `debug` to `v4.4.1` to prevent downloading the malicious `v4.4.2` package.

What happens is consumers get Puppeteer and end up downloading the bad version because of the carrot `^`. In general, it's best-practice to pin versions in `package.json` even if using Yarn or `package.lock` because consumers only get `package.json`.