Set large height explicitly for `test/assets/input/touchscreen.html`
The test page(`touchscreen.html`) was setting `touch-action: none`, but
due to a bug in chromium the page seems to be starting a gesture scroll.
Setting height explicitly seems to circumvent this bug, and a gesture
scroll is never started for such a page.

The page now receives extra pointer events since earlier a pointer
cancel was being generated, which stopped rest of the pointer sequence
to be seen by the webpage.